Move sim_setup to module subfolder
Also, now import __all__ of para_temp_setup to sim_setup (only three
functions)
